Use ecmascript on client as well
The client code actually has ES6 constructs such as `let`, `=>` and
`...args`, though most of it can be removed, the `...args` are not a
trivial change.  So the less painful option of using ecmascript on the
server side will solve the issue.

Fixes https://github.com/netanelgilad/angular-meteor-server/issues/15 by
making sure the minifiers work.